qx.ui.component.DateChooserButton is a new widget which creates a button which on execute opens a window with a date chooser.
By using the property targetWidget a target widget is associated (connected) to
the date chooser button where the selected date is synchronized with. The
target widget must be of type qx.ui.core.Widget.

The main usage would be to have a textfield which has a date chooser button
behind in order to choose a date for the textfield.

Two demos are included which show the usage. 

The first demo DateChooserButton_1.html shows the default date formats.
The second demo DateChooserButton_2.html shows a german localized version.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@5173 61aaf1ca-a10e-0410-b9fb-ac63ec25887f