Fixed memleak. The references in the object db were simply set to null, which is faster than using the delete operator but which doesn't free the hashCode!
Since JavaScript is internally storing the keys as strings, simply moving the mouse generated hundreds of strings that were all referenced by the object db and never available for garbage collection (except on page unload).

The new mechanism works via a flag that stores whether the whole application is being disposed (on page unload) or only a single object. In case of the application, the object db entries are set to null (as before), so it should be as fast as before. When only a single object is disposed, the key in the object db is deleted (thus making it eligible for garbage collection).

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@5285 61aaf1ca-a10e-0410-b9fb-ac63ec25887f