This virtual tree widget is actually starting to become relatively usable. :-)
Here, we add events generated when a tree node is contracted (treeClose),
expanded while empty (treeOpenWhileEmpty), and expanded when it already has
content (treeOpenWithContent).

In the new test program, we add a new leaf each time a branch with content is
expanded, to simulate dynamically adding the contents of the branch as the
branch is expanded.


git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@5696 61aaf1ca-a10e-0410-b9fb-ac63ec25887f