- Add a new type of pageview called RadioView.  This view provides a   presentation for asking a question and accepting one of a few pre-determined   choice answers.  Each answer then presents its own page for further options   or processing.
- New example RadioView_1.html demonstrates usage.

Note that RadioView does not provide all of the top/left/right/bottom options
of other views.  Although it could certainly be done, this seems to be an
inappropriate button-bar presentation to place other than at the top, so I
kept the implementation simple and provided only that option.

Derrell


git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@6918 61aaf1ca-a10e-0410-b9fb-ac63ec25887f