Change object db from Array to Map. I think this may lead to some discussion so I explain the benefits of doing so:
 - The hash code is always the same as the db key
 - we need to store one key less in each object (the '__dbKey' member)
 - we are able to resolve hash codes to object instances
 
The last point is important because this enables us to create "weak references" to objects. Instead of storing the object, we can store the hash code and resolve at run time. The object can than safely disposed and no dangling reference stops the GC from removing the object from memory. The hash code can be used as well to avoid circular references between the DOM and JS if only the hash code (which is an integer) is stored on the DOM element.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@11753 61aaf1ca-a10e-0410-b9fb-ac63ec25887f