generateResourceInfoCode() is now scanning the *source* locations of the libs, and is thus *independent* from any resources being copied to the build path beforehand. the paths are still generated for the build tree. to that end, generateLibInfoCode() maps namespaces to the build path in the build version (using 'forceUri' param).
add 'root' key to all "compile-source/dist" triggers. this is still inconsistent, especially with some keys in 
the compile-dist map, which has to be cleaned up. also, the 'root' value is faked for the demos in demobrowser
(in tmpl.json), since it is only used to compute the number of '..' for a relative path. this is a hack, and we
should consider allowing both a path to the app root (with 'root') and a (partial) uri (with the 'resourceUri' or
a variant of it), as alternative ways to specify the location of the HTML files relative to the lib path.


git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@13861 61aaf1ca-a10e-0410-b9fb-ac63ec25887f