add a lot of code for Job class and using code (still inactive);
background: due to config includes, it is getting increasingly difficult to maintain connection
between a job and its correct context (important for 'extend', 'run' features). due to the adding
of imported jobs to the local jobs map and the merging of job data during 'extend' resolution, this
connection gets lost. to make sure the connection remains intact, i either have to rename job
references in keys (using a scheme that allows to re-establish the connection with original config),
or replace job names with object references. i think the latter will be more clean.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@14078 61aaf1ca-a10e-0410-b9fb-ac63ec25887f