[FIX]
Fixed problem with enabling/disabling a widget with an associated command. The enabling/disabling of a parent widget affect only widgets with no command linked. For every widget with a command associated only the command was enabled/disabled not the widget itself.

Thanks to Matthew Gregory for reporting this bug!

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@15739 61aaf1ca-a10e-0410-b9fb-ac63ec25887f