[BUG #1102] this is actually a more general patch, but was discovered in the context of the 'test-source' job. it's a re-work of the handling of relative paths in general and of the 'uri' member of a library in particular. eventually, this config member shall be obsolete, since the individual library is essentially unable to know which uri needs to be used to access its resources. so the property is misplaced there.
rather, the generator can calculate the uri to the library, given the path to its Manifest file and the path to the application's
index.html. this is what the 'compile-source' job encodes in its 'root' member, and the compile-source job is also the place
that needs to handle all this information.

currently, the handling is spread over multiple places (Config handler, scanLibrary, runSouce), where parts of the final uri are
manufactured and eventually put into place (absolutizing and relativizing paths along the way). one of the reasons is the early
run of scanLibrary, where uri information is already put into the list of classes, way before runSource can evaluate its 
'compile-source/root' config entry. all of this is hard to track, but i hope to come up with a more compact solution later, 
probably together with an overhaul of the scanLibrary usage.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@15743 61aaf1ca-a10e-0410-b9fb-ac63ec25887f