[BUG #ASSIGNED 1132]: Inline window stops dragging when it gets moved too quickly http://bugzilla.qooxdoo.org/show_bug.cgi?id=1132
The "qx.ui.core.EventHandler" took a wrong target as input and found no widget instance connected to this target element. Setting the target to the captured element inside the "MouseCapture" dispatcher class solves this problem as the "qx.ui.core.EventHandler" then operates on an element which is connected to a widget.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@16382 61aaf1ca-a10e-0410-b9fb-ac63ec25887f