[Bug 605 & Bug 2021] Double click and context menu issues in TreeVirtual
TreeVirtual has always set the Table property 'alwaysUpdateCells' to true. It
was recently discovered, though, that doing so causes nodes to be removed from
the dom immediately upon a single left or right click, preventing double click
events from working in IE, and causing unexpected lack of selection even in
FF.

To solve this problem, the explicit setting of 'alwaysUpdateCells' has been
removed (leaving it to Table's default of false). The API documentation has
been annotated to reference the change of behavior. Requesting a unique icon
for display when a node is selected (in TreeVirtual) no longer works by
default. To enable that feature, the user must explicitly set the
'alwaysUpdateCells' property to true, and incur the penalty of the other
broken behavior.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@18438 61aaf1ca-a10e-0410-b9fb-ac63ec25887f