[BUG #ASSIGNED 2112]: Inline TextFields open Popup Tables that don't close upon loss of focus http://bugzilla.qooxdoo.org/show_bug.cgi?id=2112
Modified listener to support inline applications within the popup manager. Now listening to the "document.documentElement" instead of application root widget because this is not applicable for inline applications.

This solution is working but far from optimal. The goal has to be that the ui layer is *not* working with DOM elements in any way. We need to abstract this in a lower layer.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@18493 61aaf1ca-a10e-0410-b9fb-ac63ec25887f