[BUG #ASSIGNED 2185]: IE8: Performance issues with decorators http://bugzilla.qooxdoo.org/show_bug.cgi?id=2185
Enabled the use of the alpha image loader for IE8. Along with the alpha image loader a different DOM structure (7 DIV elements instead of 4 DIV and 3 IMG elements) is used. This fixes the performance issues in IE8 running in "IE8 Standards Mode".

This fix only applies for PNG images - which are exclusively used for framework throughout the framework.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@18639 61aaf1ca-a10e-0410-b9fb-ac63ec25887f