[BUG #ASSIGNED 1157]: Implement support for native windows http://bugzilla.qooxdoo.org/show_bug.cgi?id=1157
Re-Introducing native window support. This class is intentionally static to streamline it with the other classes in the "qx.bom" package.

Support for native modal windows is skipped because there is no cross-browser way to do it. Although IE, Gecko and Webkit are supporting the "showModalDialog" method, but Opera doesn't.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@18858 61aaf1ca-a10e-0410-b9fb-ac63ec25887f