[BUG #ASSIGNED 2146]: Drag&Drop Reordering broken http://bugzilla.qooxdoo.org/show_bug.cgi?id=2146
Fixed the example by making the indicator droppable (to avoid the wrong cursor when dragged over) and added some logic if the user drops directly at the indicator.

Changing the behaviour to exclude the indicator doesn't work because we need to handle the use case of dropping directly at the indicator anyway.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@18871 61aaf1ca-a10e-0410-b9fb-ac63ec25887f