[FIX BUG #2553] qx.ui.table.model.Simple.setSortMethods()
Allow passing either a map containing both an ascending sort function and a
descending sort function, or just an ascending sort function (not in a
map). The former provides backward compatibility and should still be used for
large tables (as now documented). The latter may be used as a simplified
interface. When only an ascending sort function is provided, the descending
sort function is dynamically created by calling the ascending sort function
with swapped parameters.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@19606 61aaf1ca-a10e-0410-b9fb-ac63ec25887f