[Bug #2731] Table's TOGGLE selection model does not work consistently
Although TOGGLE had been added as a model option, the selection manager had
not been modified to set the selection in the model in cases of toggle. Adding
this feature to the model required removing a generic check for whether a row
was already selected, so that check had to be moved into the SINGLE_SELECTION
case within the model.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@19935 61aaf1ca-a10e-0410-b9fb-ac63ec25887f