[Bug 2833] Rpc.refreshSession() does not refresh session information
Applied the fix suggested by Andreas J:

------- Comment #2 from andreas.junghans@stz-ida.de  2009-09-28 19:44 -------
The fix is very simple: Just remove the "if (this.getCrossDomain())" in the
refreshSession method (in the Rpc client class). The code inside the "if"
should always be executed, no matter whether the request is cross-domain.

I don't believe I have commit rights for qooxdoo-contrib, and even if I did,
this would have to wait (too much work on my plate). So it would be great if
someone else could make the mentioned change and commit it.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@20219 61aaf1ca-a10e-0410-b9fb-ac63ec25887f