[BUG #2686 FIXED]: The disappear event never gets fired http://bugzilla.qooxdoo.org/show_bug.cgi?id= 2686
registerEvent() and unregisterEvent() did not check the type of the event and only use the hash code of the target to insert the target into the target list. So unregisterEvent() did remove both events at the same time!
Now both methods build the has code with the target hash code and the event type.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@20670 61aaf1ca-a10e-0410-b9fb-ac63ec25887f