[Bug 3684] Make qx.io.remote.Rpc request object and paramters be configurable
It is now possible to extend qx.io.remote.Rpc and override the two new methods
createRequest() and createRpcData(). createRequest() instantiates the
qx.io.remote.Request() object, so an extending class can provide different
parameters to the creation of that object. createRpcData() builds the object
which is passed to the JSON-RPC server. An extending class can override this
method to issue requests in a style other than qooxdoo's default.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@22228 61aaf1ca-a10e-0410-b9fb-ac63ec25887f