[Bug 3733] Table: hover effect for rows
Rather than adding a specific effect, I've added a generic feature to allow
the developer to add attributes to each row. The default row renderer now
provides a getRowAttributes() method which may be overridden in a subclass to
add additional attributes, e.g. onmouseover and onmouseout, to provide effects
such as hover.

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@22257 61aaf1ca-a10e-0410-b9fb-ac63ec25887f