[BUG 3621 ASSIGNED] "id" of qx.bom.Collection does return wrong values when used within a chain http://bugzilla.qooxdoo.org/show_bug.cgi?id=3621
Fixed the wrong behaviour of the selector engine if the user does search for an element which does not exist. Instead of falling back to the document root node an empty collection is returned. This is especially important whenever chaining is used.

Fixed both in branch 1.1.x and trunk.
Added unit test which documents the right behaviour.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@22473 61aaf1ca-a10e-0410-b9fb-ac63ec25887f