[Bug 3789] Single click on table cell triggers cellClick event twice
What a mess. Operating systems and vendors should agree on an order of events,
to prevent developers from having to deal with this crap.

Since the order of events upon a click can vary by browser (see bug #3502), we
need to issue events in various places to maintain the qooxdoo consistent
ordering of events... which means that we also need to track whether we've
already issued an event to avoid issuing it again, later. Such was the case
here, with inadequate checks to prevent the double issuance.

This should really be tested in various browsers and on various operating
systems to ensure that click events and contextmenu events (and
MTableContextMenu) all still work properly. It _looks_ like this should only
fix the current problem and not break anything else, but...

Derrell

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@22535 61aaf1ca-a10e-0410-b9fb-ac63ec25887f