[Bug #2409]  - Problem in qx.util.StringSplit.split
The problem was the limit parameter. If used without the limit parameter,
the split method works as expected.

It seems that the StringSplit uses and older version of Steven Levithan's 
cross-browser split (0.1).  In the meantime, the limit parameter was fixed,
and the script refactored and optimized. So, I've kept using the Steven's split
method but updated it to the newest version (1.0.1). 
http://blog.stevenlevithan.com/archives/cross-browser-split

I've also implemented his demo tests as unit tests. 
http://stevenlevithan.com/demo/split.cfm



git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@22663 61aaf1ca-a10e-0410-b9fb-ac63ec25887f