[BUG 4166] Update error handling in RequestQueue class http://bugzilla.qooxdoo.org/show_bug.cgi?id=4166
Updated error handling of the RequestQueue to use the updated API of the "error" method. This way we let the Logger class handle the error object as incoming argument, creating a stacktrace and format it accordingly. This is better than let each class creating the stacktrace from an error object.

Fixed for both trunk and branch 1_2_x

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/branches/branch_1_2_x@23468 61aaf1ca-a10e-0410-b9fb-ac63ec25887f