[BUG 3512] Optimizations at "qx.dom.Hierarchy" broke Iframe loading under Webkit http://bugzilla.qooxdoo.org/show_bug.cgi?id=3512
Analyzed the initial commit which broke the iframe loading under Webkit. The issue here was that the "isRendered" method did not check for elements which are not in the DOM.
Unfortunately, the "contains" method at the BODY element does return "true" for an element which is *not* in the DOM yet. Adding extra code to check if the DOM element is not in the DOM solves this issue.

Added unit tests to prevent any regression bugs in the future.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@23607 61aaf1ca-a10e-0410-b9fb-ac63ec25887f