[BUG #4607] added test to assert behavior when resource is loaded from cache
Verify that browser does progress readyState to DONE
even when resource was loaded from cache. According to jQuery,
there is a bug in some versions of IE. However, this test
does not fail in the versions of IE I have tested so far.

According to Fiddler, the resource is indeed loaded from cache
in the second request.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@25761 61aaf1ca-a10e-0410-b9fb-ac63ec25887f