[BUG #4607] Basic authentication implemented as IAuthDelegate
Note: Setting username and password in the header is more secure than passing credentials to the transports open method, because the latter adds sensitive information to the URL (which may be logged). Therefore, qx.io.request.Xhr does not provide a measure pass the credentials this way. Basic auth as defined in the delegate should be used instead.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@26894 61aaf1ca-a10e-0410-b9fb-ac63ec25887f