[BUG #5104] Use new transport by default
After some discussion, we decided to change the default. This makes it possible to deprecate old code in one release (instead of two) and avoids the introduction of another constructor param.

Users who want to use the old transport need to explicitly set deprecatedTransport. Added upgrade notice to API docs.

git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@27312 61aaf1ca-a10e-0410-b9fb-ac63ec25887f