[Bug 5789] qx.ui.container.SlideBar prevents page scrolling with mouse wheel
The mouse wheel has two axes in which it can move. It can rotate, which is
intended to invoke a vertical scroll. On some mice, it can also press
sideways, which is intended to invoke a horizontal scroll. The rotation should
never invoke a horizontal scroll, and the press side-to-side should never
invoke a vertical scroll.

This commit adds a getAxis() method to qx.event.type.MouseWheel which is then
used in ScrollBar to cause the mouse wheel to work properly on that widget.

Derrell