[BUG #4904] Move styles common to both themes to shared partial
The idea is to collect style settings necessary for a usable layout,
while moving all theme specific properties (colors, borders, fancy eye
candy) to the corresponding theme stylesheet. This way layout and
theming can be better separated.

The widgets partial can be used as skeleton for new themes. For common
selectors found in the theme files, stubs are provided (they are
stripped by less if no style properties are assigned).