Load scripts in parallel, but execute in order
Experimental feature. Append ?parallel to URL to enable.

Modern browsers try to load scripts in parallel whenever possible
and delay execution until all scripts are available. Traditionally,
scripts inserted in HTML markup were blocking (synchronous), i.e.
delayed loading of *all* subsequent resources. A common work-around
is to insert script elements to the DOM, and thats exactly the
approach the script loader takes. To ensure that scripts are executed
in order, each script is inserted once the previous script was loaded.

a---
    b---
        c---

The draw back of this approach is that only one connection to the
server is used for loading scripts.

This patch enables parallel loading by appending all scripts at once,
but marking them synchronous.

a---
b---
c---

Basically this mimicks the behavior of the browser when in encounters
a list of external script elements in the HTML head. Browsers not supporting
the "async" attribute will probably execute scripts in the order they arrive.