[Bug 6092] Inappropriate messages in console when using data URLs (wrong fix)
The fix for this bug was broken. It allowed the possibility to dereference a
null or undefined value when determining the width and height of an
image. This patch fixes that problem, using the configured default width and
height for an image if not available from the image data itself.

Derrell