[Bug 6175] qx.ui.core.scroll.MWheelHandling prevents scroll Y when only X scrollbar
This is the same problem as we solved for SlideBar in bug 5789, but this one
effects qx.ui.container.Scroll via its use of mixin MWheelHandling.

The problem is that if the scroll container has only a horizontal scrollbar,
using the mousewheel to scroll vertically beyond the scroll container is
blocked. The solution is very similar to that used to fix bug 5789: check the
deltaX and deltaY values and don't stop propagation if the relevant scroll bar
doesn't exist for the container.

Derrell