Allow paths/approot to be null.
The user then has to specify urls for all libraries manually.

This is particularily helpful when the main config file is in memory
and has no base path, because in those cases the default paths/approot
value from Qooxdoo's base json files, which is relative, won't work for
lack of a base path.

However, currenlty, with no paths/approot set, generating the uri for the
implicit __out__ library still causes trouble. See:

	http://bugzilla.qooxdoo.org/show_bug.cgi?id=6616