Allow paths/approot to be null.
The user then has to specify urls for all libraries manually.

This is needed to support config files from memory without a base path,
in conjunction with the Qooxdoo base build config files, which use a
relative default for paths/app-root, and so this particular call would
result in an error, because said path value cannot be made absolute.