Added null check to Spinner.js which fixes the following bug:
The property numberFormat of qx.ui.form.Spinner is nullable. If you do something like

spinner.setNumberFormat(new qx.util.format.NumberFormat());
spinner.setNumberformat(null);

it will produce an error in the shape of:

/framework/source/class/qx/ui/form/Spinner.js:498 Uncaught TypeError: Cannot read property 'addListener' of null

because it tries

this.getNumberFormat().addListener(...);

although this.getNumberFormat() is null.

Fixed it with an "if (this.getNumberFormat() !== null)".