Fixing qx.util.format.DateFormat test on locale reset
I suspect that this test did not work for long time, as resetting the
locale allways lead to a null value exception in the bound DateFormat
instance (var df).

Now we create a DateFormat instance which uses the initial locale and
use it to complare against the reset bound DateFormat.