Issue #9303: Improve mouse wheel normalization code to adjust scrolling sensitivity for touchpads (#9306)
* #9303 : Improve mouse wheel normalization code to adjust scrolling sensitivity across browsers

* #9303 :
- Touch pad detection limited to Mac OS after Sierra (>= 10.12).
- Events occurring during detection are now buffered and scheduled for
fire after the detection period ends if no other wheel events come. This
will prevent single wheel events to be swallowed and physical wheels
to seem broken.
- If a touch pad is detected, the delta normalization code is ignored
and a factor of 1 is applied to the delta of the DOM event. This lets
the native delta computation happen.
- Constants have been reviewed. The TOUCHPAD_WHEEL_EVENTS_TIMEOUT
has been set to 5s in order to trigger a new detection process.
- A new qx.core.Environment entry (qx.debug.touchpad.detection) has been
 added to print touchpad debug messages. It is set to false by default.

* #9303 : Minor fixes after the previous PR review