Fix Textarea autosize calculation in special case (#9321)
* Fix Textarea autosize calculation in special case

In some cases the textarea's autosize method was called before the textarea had proper bounds applied. In this case a clone of the textarea was created without a width set and the width was also not updated later. This lead to a wrong calculation of textarea height, because the clone used for calculation had a very small width, while the real textarea could be much bigger, which produced a lot of whitespace the longer the text got.

This fix delays the creation of the clone until the bounds are available. Fixed the issue in our project.

* Fix textarea autosize when element size changes

Initial fix reverted (caused problems and did not fully solve the problem) and using another approach now, which also covers resizing browser window.

* Fix textarea autosize when element size changes 

Initial fix reverted (caused problems and did not fully solve the problem) and using another approach now, which also covers resizing browser window. (sorry last commit was based on wrong file version)