Allow RadioGroup to group by other than the 'value' property.
Prior to this change, widgets to be grouped in a RadioGroup had to have a
'value' property which was monitored for changes to ensure that no more than
one of the grouped widgets was active at any one time; and a 'group'property,
which was set to the RadioGroup object to which the widget was added.

With this change, the two property names can be specified. The RadioGroup
object has two new properties:

- The property groupedProperty, which defaults to 'value', is the property on
  which widgets are grouped.

- The property groupProperty, which defaults to 'group', is the property in
  the widgets which is set by RadioGroup to tell a widget what RadioGroup
  object it is associated with.

It is now possible for a set of widgets to have multiple properties that are
independently maintained as mutually exclusive via multiple RadioGroups, each
responsible for the mutual exclusivity of a particular property. For example,
if a group of widgets can have only one of them with their 'locked' property
set to true, and the same set of widgets can have only one of them with their
'ready' property set to true, then the widgets would be defined with
properties:

    locked :
    {
      init     : null,
      check    : "Boolean",
      apply    : "_applyLocked",
      event    : "changeLocked",
      nullable : false
    },

    // set by the RadioGroup
    lockedGroup :
    {
      init     : null,
      check    : "qx.ui.form.RadioGroup",
      nullable : true
    },

    ready :
    {
      init     : false,
      check    : "Boolean",
      apply    : "_applyReady",
      event    : "changeReady",
      nullable : false
    },

    // set by the RadioGroup
    readyGroup :
    {
      init     : null,
      check    : "qx.ui.form.RadioGroup",
      nullable : true
    },

Then two radio groups can be defined:

    var rgLocked = new qx.ui.form.RadioGroup();
    var rgReady = new qx.ui.form.RadioGroup();

    rgLocked.set(
      {
        groupedProperty     : "locked",
        groupProperty       : "lockedGroup"
      });

    rgReady.set(
      {
        groupedProperty     : "ready",
        groupProperty       : "readyGroup"
      });

Then, widgets which define the 'locked', 'lockedGroup', 'ready', and
'readyGroup' properties can be added to rgLocked and rgReady.