Remove use of qx.bom.Selector in qx.util.ResourceManager
ResourceManager is used in non-GUI code. qx.bom.Selector
brings in the Sizzle library, which does a whole bunch of
initialization that requires a GUI environment (expects
'document' to exist).

This patch replaces the use of qx.bom.Selector with explicit
tests for existence of 'window' and 'window.document' and
use of the element API.