sass support is working
I had to change the api of the manager. It's now constructed with a reference to the used analyser. So the libraries collection could be used in both and i must'nt copy the code for finding the qooxdoo library.
We could think about an LibraryList interface in place of the whole analyser.