Catch false positive parsing of date strings via Date.parse in safari
Check if parsing a date string via the original Date.parse method returns a real time value, which we ensure by using setTime with an intermediate Date object and the parsed time value. 

Safari 11 e.g. parses the date string '19700101' successfully into a time value, but returns NaN if that value is used in setTime.

See:
  https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/parse
  https://github.com/qooxdoo/qooxdoo/issues/9451