Fix webfont rule removal for IE and edge
In IE and edge even if the rule was added with font-style first and font-weight second, it is not guaranteed that the attributes remain in that order. Therefore we check for both versions, style first, weight second and weight first, style second.
Without this fix the rule isn't found and removed reliable. 

This also makes the test qx.test.bom.webfonts.Manager."test: do not create rule for invalid font" succeed again on IE and edge