Testrunner: enqueue qx.test.bom.media.MediaTestCase.testPlayEvent as the first test to make it succeed on safari
On safari this test fails, if we are running the complete testsuite and the test is not the first test in qx.test.bom.media.MediaTestCase. By naming it with a blank behind test it is enqueued as the first test and the test run succeeds.
 
Note that the tests in qx.test.bom.media.MediaTestCase are started seperately or this test is restarted after it failed, then it succeeds.
This is some weird timing or caching issue which I could not solve otherwise.