Enable API users to know about parser errors
If the data source is providing an invalid JSON file (for whatever
reason), you'll run into a deadlock when waiting for some kind of
response from the data store / Xhr request.

It will neither fire 'loaded', nor 'fail', nor 'error'. You can
just see an error on the browser console. The error itself is not
catchable.

Adding a 'parseError' event solves this problem.