oo loader template could fail to initialize userAgent (#9492)
* oo loader template could fail to initialize userAgent

In the React Native environment, window.navigator is defined, but
window.navigator.userAgent is not. This patch ensures that userAgent gets
set even if window.navigator was previously set. For consistency, it also
sets window.navigator.product and window.navigator.cpuClass if either is
unset, regardless of whether only those individual members were unset or if
window.navigator was originally unset.

* fix a typo