Add flush, some timeout and a longer in order to reliably succeed tests on firefox
The test "test dropdown list wider than selectbox" creates a test string ("asd.....") which in firefox has a width of ~650px where in chrome it has ~670px. therefore the test fails in firefox were the assertion expected a width of greater than 666px. Addes some more charcters to reliably fullfill the expected width. 
Both tests are not running reliable on high loads and in containers because of the virtual box needs some time to create the dom and calculate width etc. therefore a flush was added and the test "test dropdown list same width as selectbox" has now a timout of 10 before the results are evalutated