Remove global invalid tooltip on manager disposal
If a form is currently invalid and the validation manager gets disposed,
it can happen that the tooltip is still visible - even if the form and
the widget which receives the invalid message are already disposed.

Calling _showTooltip(false) (strange naming due to inverse logic, though)
does everything we need to close the tooltip on disposal.