More SelectBox updates
Added animation to make SelectBox ease and grow at the same time. Tried to make it to where it would do that going up and down. Since i have to add it at appear event it will execute the previously added animation before executing the right one. Not sure how to fix. I will leave as is and address it as a very low bug