Allow not throwing an error when non-focusable widgets are focused/blurred
In certain cases, Windows need be made unfocusable, yet internal qooxdoo code,
in the bowels of Promise-based code, tries to focus or blur them. These focus
and blur calls can be safely ignored, when nothing is expecting the focus or
blur to have an effect, yet display of the thrown error could not be
prevented.

This is done on a static, global basis. Applications that require it know that
they require it (few do, but when you need it, you really need it!), and can
change from the default, which remains as it was previously, for backward
compatability. I don't see any reason for this to be a property. Having a
property would require *also* having a way to specify the default value of the
property.