Implemented migration path
On first use of a qx package|contrib command, the user is alerted to first
run `qx package --migrate`, then to run `qx package upgrade --reinstall`.
This should fix the paths in the lockfile. The user then is alerted
to adapt files like `.gitignore` to ignore the new "package" dir. This
could of course be done automatically, too.