Fixes #8915 - drop on iFrame
This fixes the problem, but I don't understand how a drop is a mutation, so
I'm not sure why it was failing previously. This was modified to add
MutationObserver by @schnepe2 in commit 1086fad.

The tests are disabled for this class, too. I'm not sure how to confirm that
this change doesn't break anything else...