Fixes #8915 - drop on iFrame (#9838)
* Allow changing the header cell renderer of many columns without a large delay

This fixes #9548 by adding a new method, setHeaderCellRenderers, which fires
the "headerCellRendererChanged" event only once after setting all renderers,
as opposed to a user program calling setHeaderCellRenderer many times, once
for each renderer change, which causes "headerCellRendererChanged" to be fired
after each change. The processing done internally by the framework upon that
event is slow when run multiple times, and there is no need internally for it
to run more than once after all of the renderers have been changed.

* Fixes #8915 - drop on iFrame

This fixes the problem, but I don't understand how a drop is a mutation, so
I'm not sure why it was failing previously. This was modified to add
MutationObserver by @schnepe2 in commit 1086fad.

The tests are disabled for this class, too. I'm not sure how to confirm that
this change doesn't break anything else...

Thanks @derrell !