Virtual list page scroll (#9837)
* Allow changing the header cell renderer of many columns without a large delay

This fixes #9548 by adding a new method, setHeaderCellRenderers, which fires
the "headerCellRendererChanged" event only once after setting all renderers,
as opposed to a user program calling setHeaderCellRenderer many times, once
for each renderer change, which causes "headerCellRendererChanged" to be fired
after each change. The processing done internally by the framework upon that
event is slow when run multiple times, and there is no need internally for it
to run more than once after all of the renderers have been changed.

* Enable PageUp and PageDown to operate by pages. (Was never really implemented)

PageUp and PageDown were "implemented" as moving to the first or last
selectable item in the list. This implements proper full-pane paging when
PageUp or PageDown are pressed.


Thanks @derrell !