Avoid flickering in variable height qx.ui.list.List
When having the property `variableRowHeight` set to true (which is the default), the qx.ui.list.List items flicker noticeably when manipulating the model of the list (e.g. be removing an item).
 
This is caused by unnecessarily resetting all item heights in method `__updateGroupRowHeight` (shrink all rows to default) and setting them again later on in `_setRowItemSize`  (set row height from item height) is called deferred.

This PR avoids this by leaving `__updateGroupRowHeight` when  `variableRowHeight` is true.