Fix jsdoc parameter type of MProperty.prototype.set [ci skip ci]
The the type of the first parameter of `MProperty.prototype.set` is not `Map | String`, but `Object | String` since `Map` is very specific type (https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Map) not an `Object` *used* as a map or dictionary. We might have this documentation problem at other places, too, but it is specifically salient here because it leads to linter warnings in IDEs.