Trailing rows could be left hanging in rendered trees
If rows immediately prior to the last row were removed with the table
model's `removeRows` method, the final row(s) would not get updated,
leaving stale and non-interactive rows at the end of the rendered
table. This Fixes #10365.

The above change broke removing the last row(s) of a table as a result
of a long-time bug, also fixed herein.