Initial cut at a simple implementation of browserifying commonjs modules
This implements a working proof of concept. It requires that the
compile.json target indicates `application-types : [ "browser" ]`
exclusively (i.e., it doesn't yet build browser and node apps
separately, bringing in the browserified files only in the browser
version.

Usage:
In a target, specify the application type as exclusively "browser",
and add an array of CommonJS modules to be browserified. For example:
```
      "application-types": [
        "browser"
      ],
      "commonjs-modules": [
        "semver"
      ],
```

You may also add a similar `commonjs-modules` array in an
`application` entry, in which case those CommonJS modules will be
browserified for all targets which indicate "browser" as the only
application type. The resulting set of modules to be browserified is
the union of those indicated in `applciations` and those indicated in
the `target`.