Automatically detect and browserify CommonJS modules (#10378)
* Initial cut at a simple implementation of browserifying commonjs modules

This implements a working proof of concept. It requires that the
compile.json target indicates `application-types : [ "browser" ]`
exclusively (i.e., it doesn't yet build browser and node apps
separately, bringing in the browserified files only in the browser
version.

Usage:
In a target, specify the application type as exclusively "browser",
and add an array of CommonJS modules to be browserified. For example:
```
      "application-types": [
        "browser"
      ],
      "commonjs-modules": [
        "semver"
      ],
```

You may also add a similar `commonjs-modules` array in an
`application` entry, in which case those CommonJS modules will be
browserified for all targets which indicate "browser" as the only
application type. The resulting set of modules to be browserified is
the union of those indicated in `applciations` and those indicated in
the `target`.

* Allow multiple application types in target as long as "browser" is present

* Move browserifying to compiler.targets.meta.Browserify

* Completd auto-detection of `require`d modules; now stored in database

* Use browserify API; show module name and throw when there is a  missing module

* Track locations of require() to display in error if module not found

* Make builtins (including `process`) work better

* Lint is configured to be way, way overly picky!

* More stupid lint complaints

* `require` is no longer unresolved

* Update release notes

* Add documentation for using `require`

* Remove throw upon unfound module. Hopefully error message is obvious.

Thanks @derrell !