Convert `import` statements to `require()` in `require()`d dependencies
This applies to the automatically discovered use of `require()` in
browser apps, that causes those `require()`d modules to be bundled and
included in the app. The problem was that `import` statements need to
be at top level, but due to the wrapping generated by the qooxdoo
compiler, they ended up not at the top level.

This adds use of browserify plugin `esmify` that converts `import` to
`require()`.