Click order fix (#10632)
* Update Button.js

It seems that touching a button (at least on ios and android)
does not immediately focus it before triggering the tap event
this causes problem with change events for input fields that
previously held focus, as they will arrive AFTER the button has
already executed.

* Update AbstractField.js

This is not necessary anymore as the focus changes early enough

* Update Button.js

switch to using native events

* Update Button.js

the event handler is fine ... the problem is just the 'focus' call .. that needs to be on the DOM element

* Update source/class/qx/ui/form/Button.js

Co-authored-by: Dmitry Zolotov <goldim1993@mail.ru>

---------

Co-authored-by: Henner Kollmann <henner.kollmann@gmx.de>
Co-authored-by: Dmitry Zolotov <goldim1993@mail.ru>

Thanks @oetiker!