Placeholder text does not get correct theme color.
Code in ui/form/AbstractField.js looks broken for some time to me (11th March 2013).
__stylesheet was replaced on most places by using "global stylesheet" code.

But the _onChangeTheme() function still contains this if statement:
if (!this.__useQxPlaceholder && qx.ui.form.AbstractField.__stylesheet) (which is not used any more)
So the color from the loaded theme is never applied.

Now using the theme name to check when an update is needed.

See https://github.com/qooxdoo/qooxdoo/issues/10694